+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Fixed
+- Fix generate associated item massive action
+
 ## [2.12.6] - 2026-02-23
+
+### Fixed
 
 - Fix SQL query error when displaying deliveries
 - Fix error during generate associated material action