+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [unreleased]
 
 ### Fixed
+
 - Fix generate associated item massive action
+- Update locales
+
 
 ## [2.12.6] - 2026-02-23