+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix generate associated item massive action
 - Update locales
-
+- Migration of reports to next version of plugin Reports
 
 ## [2.12.6] - 2026-02-23