+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix generate associated item massive action
+- Prevent invalid references and non-existent classes during generate item flow.
 - Update locales