+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix generate associated item massive action
 - Prevent invalid references and non-existent classes during generate item flow.
 - Update locales
+- Fix `glpi_plugin_order_referencefrees` table not removed on uninstall
 
 
 ## [2.12.6] - 2026-02-23