+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Prevent invalid references and non-existent classes during generate item flow.
 - Remove unnecessary message during plugin installation
 - Update locales
+- Fix getDatabaseRelations() declarations to fix warnings on order deletion
 - Fix `glpi_plugin_order_referencefrees` table not removed on uninstall