+++ CHANGELOG.md
@@ -5,10 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.12.7] - 2026-06-03
+## [2.12.8] - 2026-06-24
 
 ### Added
 - Add an `Orderable` capacity for GLPI 11 custom asset definitions.
+
+
+## [2.12.7] - 2026-06-03
 
 ### Fixed